Revert "Refactoring of initial sync (#5096)"
This reverts commit 3ec2a0f9e0924b9ce06159916e89742a85521033.
