+++ beacon-chain/node/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//beacon-chain/state/stategen:go_default_library",
         "//beacon-chain/sync:go_default_library",
         "//beacon-chain/sync/initial-sync:go_default_library",
-        "//beacon-chain/sync/initial-sync-old:go_default_library",
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
