+++ beacon-chain/node/node.go
@@ -34,7 +34,6 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/stategen"
 	prysmsync "github.com/prysmaticlabs/prysm/beacon-chain/sync"
 	initialsync "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync"
-	initialsyncold "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync-old"
 	"github.com/prysmaticlabs/prysm/shared"
 	"github.com/prysmaticlabs/prysm/shared/cmd"
 	"github.com/prysmaticlabs/prysm/shared/debug"
@@ -405,19 +404,9 @@ func (b *BeaconNode) registerSyncService(ctx *cli.Context) error {
 		return err
 	}
 
-	var initSync prysmsync.Checker
-	if cfg := featureconfig.Get(); cfg.EnableInitSyncQueue {
-		var initSyncTmp *initialsync.Service
-		if err := b.services.FetchService(&initSyncTmp); err != nil {
-			return err
-		}
-		initSync = initSyncTmp
-	} else {
-		var initSyncTmp *initialsyncold.Service
-		if err := b.services.FetchService(&initSyncTmp); err != nil {
-			return err
-		}
-		initSync = initSyncTmp
+	var initSync *initialsync.Service
+	if err := b.services.FetchService(&initSync); err != nil {
+		return err
 	}
 
 	rs := prysmsync.NewRegularSync(&prysmsync.Config{
@@ -442,25 +431,16 @@ func (b *BeaconNode) registerInitialSyncService(ctx *cli.Context) error {
 		return err
 	}
 
-	if cfg := featureconfig.Get(); cfg.EnableInitSyncQueue {
-		is := initialsync.NewInitialSync(&initialsync.Config{
-			DB:            b.db,
-			Chain:         chainService,
-			P2P:           b.fetchP2P(ctx),
-			StateNotifier: b,
-			BlockNotifier: b,
-		})
-		return b.services.RegisterService(is)
-	}
-
-	is := initialsyncold.NewInitialSync(&initialsyncold.Config{
+	is := initialsync.NewInitialSync(&initialsync.Config{
 		DB:            b.db,
 		Chain:         chainService,
 		P2P:           b.fetchP2P(ctx),
 		StateNotifier: b,
 		BlockNotifier: b,
 	})
+
 	return b.services.RegisterService(is)
+
 }
 
 func (b *BeaconNode) registerRPCService(ctx *cli.Context) error {
@@ -474,19 +454,9 @@ func (b *BeaconNode) registerRPCService(ctx *cli.Context) error {
 		return err
 	}
 
-	var syncService prysmsync.Checker
-	if cfg := featureconfig.Get(); cfg.EnableInitSyncQueue {
-		var initSyncTmp *initialsync.Service
-		if err := b.services.FetchService(&initSyncTmp); err != nil {
-			return err
-		}
-		syncService = initSyncTmp
-	} else {
-		var initSyncTmp *initialsyncold.Service
-		if err := b.services.FetchService(&initSyncTmp); err != nil {
-			return err
-		}
-		syncService = initSyncTmp
+	var syncService *initialsync.Service
+	if err := b.services.FetchService(&syncService); err != nil {
+		return err
 	}
 
 	genesisValidators := ctx.GlobalUint64(flags.InteropNumValidatorsFlag.Name)
