--- beacon-chain/sync/initial-sync-old/BUILD.bazel
@@ -1,63 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "log.go",
-        "round_robin.go",
-        "service.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync-old",
-    visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//beacon-chain/blockchain:go_default_library",
-        "//beacon-chain/core/feed:go_default_library",
-        "//beacon-chain/core/feed/block:go_default_library",
-        "//beacon-chain/core/feed/state:go_default_library",
-        "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/db:go_default_library",
-        "//beacon-chain/flags:go_default_library",
-        "//beacon-chain/p2p:go_default_library",
-        "//beacon-chain/sync:go_default_library",
-        "//proto/beacon/p2p/v1:go_default_library",
-        "//shared:go_default_library",
-        "//shared/bytesutil:go_default_library",
-        "//shared/featureconfig:go_default_library",
-        "//shared/mathutil:go_default_library",
-        "//shared/params:go_default_library",
-        "//shared/roughtime:go_default_library",
-        "@com_github_kevinms_leakybucket_go//:go_default_library",
-        "@com_github_libp2p_go_libp2p_core//peer:go_default_library",
-        "@com_github_paulbellamy_ratecounter//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["round_robin_test.go"],
-    embed = [":go_default_library"],
-    race = "on",
-    tags = ["race_on"],
-    deps = [
-        "//beacon-chain/blockchain/testing:go_default_library",
-        "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/db/testing:go_default_library",
-        "//beacon-chain/p2p/peers:go_default_library",
-        "//beacon-chain/p2p/testing:go_default_library",
-        "//beacon-chain/state:go_default_library",
-        "//beacon-chain/sync:go_default_library",
-        "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/hashutil:go_default_library",
-        "//shared/params:go_default_library",
-        "//shared/roughtime:go_default_library",
-        "//shared/sliceutil:go_default_library",
-        "@com_github_kevinms_leakybucket_go//:go_default_library",
-        "@com_github_libp2p_go_libp2p_core//network:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-    ],
-)
