--- beacon-chain/sync/initial-sync-old/log.go
@@ -1,7 +1,0 @@
-package initialsyncold
-
-import (
-	"github.com/sirupsen/logrus"
-)
-
-var log = logrus.WithField("prefix", "initial-sync")
